删除多余的信息和加操作人

+++ admin-api/src/main/java/com/atzuche/order/admin/service/log/SupplementSendmsgLogService.java
@@ -84,7 +84,7 @@ public class SupplementSendmsgLogService {
         OrderSupplementPayMq orderSupplementPayMq = new OrderSupplementPayMq();
         orderSupplementPayMq.setOrderNo(reqVO.getOrderNo()+"");
         orderSupplementPayMq.setType(reqVO.getMessageType());
-        orderSupplementPayMq.setRenterMemNo(reqVO.getMemNo());
+        orderSupplementPayMq.setRenterMemNo(reqVO.getMemNo().intValue());
         orderMessage.setMessage(orderSupplementPayMq);
         switch(platform){
             case 0 ://系统
